chore: clippy recommendations
