--- src/cli/cli.rs
@@ -1,13 +1,0 @@
-use clap::{crate_authors, crate_name, crate_version, App};
-
-use super::argument::make_arguments;
-
-/// Creates a **clap** application, bind arguments to it and then get matches
-/// from the exection
-pub fn build() -> App<'static, 'static> {
-    App::new(crate_name!())
-    .version(crate_version!())
-    .author(crate_authors!())
-    .about("Configurable and simple command-line HTTP server\nSource: https://github.com/EstebanBorai/http-server")
-    .args(&make_arguments())
-}
